app/server/requirements.txt
Flask==2.1.1
PyJWT==2.8.0
